Solve day 11

+++ Day7.cpp
@@ -1,13 +1,14 @@
 #include "Day7.h"
 
+#include <cstdint>
 #include <vector>
 #include <string>
 #include "StringUtils.h"
 
 namespace {
-	bool solutionExists(long long testVal, 
-						const std::vector<long long>& operands, 
-						long long val = 0, 
+	bool solutionExists(int64_t testVal, 
+						const std::vector<int64_t>& operands, 
+						int64_t val = 0, 
 						size_t i = 0) {
 		if (i == operands.size()) {
 			return testVal == val;
@@ -22,10 +23,10 @@ namespace {
 			   solutionExists(testVal, operands, val + operands[i], i + 1);
 	}
 
-	bool solutionExistsWithConcat(long long testVal,
-							  const std::vector<long long>& operands,
-							  long long val = 0,
-							  size_t i = 0) {
+	bool solutionExistsWithConcat(int64_t testVal,
+							      const std::vector<int64_t>& operands,
+							      int64_t val = 0,
+							      size_t i = 0) {
 		if (i == operands.size()) {
 			return testVal == val;
 		}
@@ -38,7 +39,7 @@ namespace {
 		bool result = solutionExistsWithConcat(testVal, operands, val * operands[i], i + 1) ||
 					  solutionExistsWithConcat(testVal, operands, val + operands[i], i + 1);
 		if (!result) {
-			long long concat{};
+			int64_t concat{};
 			if (tryConcat(val, operands[i], concat)) {
 				result |= solutionExistsWithConcat(testVal, operands, concat, i + 1);
 			}
@@ -49,11 +50,11 @@ namespace {
 
 void solveDay7Part1(std::istream& input, std::ostream& output) {
 	std::string line;
-	long long sum = 0;
+	int64_t sum = 0;
 	while (getline(input, line)) {
 		auto eq = split(line, ": ");
-		long long testVal = std::stoll(eq[0]);
-		std::vector<long long> operands = split<long long>(eq[1], " ", stollFunc);
+		int64_t testVal = std::stoll(eq[0]);
+		std::vector<int64_t> operands = split<int64_t>(eq[1], " ", stollFunc);
 		if (solutionExists(testVal, operands)) {
 			sum += testVal;
 		}
@@ -63,11 +64,11 @@ void solveDay7Part1(std::istream& input, std::ostream& output) {
 
 void solveDay7Part2(std::istream& input, std::ostream& output) {
 	std::string line;
-	long long sum = 0;
+	int64_t sum = 0;
 	while (getline(input, line)) {
 		auto eq = split(line, ": ");
-		long long testVal = std::stoll(eq[0]);
-		std::vector<long long> operands = split<long long>(eq[1], " ", stollFunc);
+		int64_t testVal = std::stoll(eq[0]);
+		std::vector<int64_t> operands = split<int64_t>(eq[1], " ", stollFunc);
 		if (solutionExistsWithConcat(testVal, operands)) {
 			sum += testVal;
 		}
